Update default credits value

+++ scripts/deploy_backend.sh
@@ -62,6 +62,24 @@ then
 	exit -1
 fi
 
+if test -z "$STRIPE_KEY"
+then
+	echo "ERROR: STRIPE_KEY variable must be defined, exiting"
+	exit -1
+fi
+
+if test -z "$SUBSCRIPTION_PRICE_ID"
+then
+	echo "ERROR: SUBSCRIPTION_PRICE_ID variable must be defined, exiting"
+	exit -1
+fi
+
+if test -z "$CREDITS_PRICE_ID"
+then
+	echo "ERROR: CREDITS_PRICE_ID variable must be defined, exiting"
+	exit -1
+fi
+
 echo "INFO: stopping already running docker container"
 ssh -o StrictHostKeyChecking=no -i key.pem azureuser@"$BACKEND_DOMAIN" "docker stop wasp-backend || echo 'No containers available to stop'"
 ssh -o StrictHostKeyChecking=no -i key.pem azureuser@"$BACKEND_DOMAIN" "docker container prune -f || echo 'No stopped containers to delete'"
@@ -75,4 +93,4 @@ echo "Deleting old image"
 ssh -o StrictHostKeyChecking=no -i key.pem azureuser@"$BACKEND_DOMAIN" "docker system prune -f || echo 'No images to delete'"
 
 echo "INFO: starting docker container"
-ssh -o StrictHostKeyChecking=no -i key.pem azureuser@"$BACKEND_DOMAIN" "docker run --name wasp-backend -p $PORT:$PORT -e PORT='$PORT' -e DATABASE_URL='$DATABASE_URL' -e WASP_WEB_CLIENT_URL='$WASP_WEB_CLIENT_URL' -e JWT_SECRET='$JWT_SECRET' -e GOOGLE_CLIENT_ID='$GOOGLE_CLIENT_ID' -e GOOGLE_CLIENT_SECRET='$GOOGLE_CLIENT_SECRET' -e AZURE_OPENAI_API_KEY='$AZURE_OPENAI_API_KEY' -e ADS_SERVER_URL='$ADS_SERVER_URL' -d ghcr.io/$GITHUB_REPOSITORY:$TAG"
+ssh -o StrictHostKeyChecking=no -i key.pem azureuser@"$BACKEND_DOMAIN" "docker run --name wasp-backend -p $PORT:$PORT -e PORT='$PORT' -e DATABASE_URL='$DATABASE_URL' -e WASP_WEB_CLIENT_URL='$WASP_WEB_CLIENT_URL' -e JWT_SECRET='$JWT_SECRET' -e GOOGLE_CLIENT_ID='$GOOGLE_CLIENT_ID' -e GOOGLE_CLIENT_SECRET='$GOOGLE_CLIENT_SECRET' -e AZURE_OPENAI_API_KEY='$AZURE_OPENAI_API_KEY' -e ADS_SERVER_URL='$ADS_SERVER_URL' -e STRIPE_KEY='$STRIPE_KEY' -e SUBSCRIPTION_PRICE_ID='$SUBSCRIPTION_PRICE_ID' -e CREDITS_PRICE_ID='$CREDITS_PRICE_ID' -d ghcr.io/$GITHUB_REPOSITORY:$TAG"
